get_tarot_reading/requirements.txt
google-generativeai
boto3
psycopg2-binary